refactor: remove transactional folder and manage react-email deps from one place
