+++ package.json
@@ -10,6 +10,7 @@
     "build": "next build && next-sitemap",
     "postbuild": "next-sitemap",
     "dev": "NODE_OPTIONS='--inspect' next dev --turbo",
+    "email": "email dev --dir ./src/utils/email/templates",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "start": "next start",
     "test": "pnpm exec playwright test",
@@ -59,6 +60,7 @@
     "react": "^18.3.1",
     "react-aria": "^3.34.1",
     "react-dom": "^18.3.1",
+    "react-email": "2.1.6",
     "react-icons": "^5.3.0",
     "react-stately": "^3.32.1",
     "rehype-autolink-headings": "^7.1.0",
