+++ src/utils/email/templates/WelcomeEmail.tsx
@@ -15,19 +15,15 @@ import {
   Text,
 } from '@react-email/components';
 
-interface WelcomeEmailProps {
+type WelcomeEmailProps = {
   firstName?: string;
-}
-
-const baseUrl = process.env.VERCEL_URL
-  ? `https://${process.env.VERCEL_URL}`
-  : '';
+};
 
 export const WelcomeEmail = ({ firstName }: WelcomeEmailProps) => {
   return (
     <Html>
       <Head />
-      <Preview>🧋 Let's build things that matter</Preview>
+      <Preview>🧋 Let&apos;s build things that matter</Preview>
 
       <Tailwind>
         <Body className="mx-auto my-auto bg-[#fafafa] px-2 font-sans text-xl">
@@ -51,14 +47,16 @@ export const WelcomeEmail = ({ firstName }: WelcomeEmailProps) => {
                   <Text style={paragraph}>Hi {firstName ?? 'friend'},</Text>
 
                   <Text style={paragraph}>
-                    Thanks for trusting me with a spot in your inbox—I'm
+                    Thanks for trusting me with a spot in your inbox—I&apos;m
                     thrilled to have you here!
                   </Text>
-                  <Text style={paragraph}>Here's what you can expect:</Text>
+                  <Text style={paragraph}>
+                    Here&apos;s what you can expect:
+                  </Text>
 
                   <ol style={{ paddingLeft: 20 }}>
                     <li style={paragraph}>
-                      <b>One Tiny idea:</b> In each dispatch, I'll share a
+                      <b>One Tiny idea:</b> In each dispatch, I&apos;ll share a
                       simple, actionable idea to help you design and build
                       better products.
                     </li>
@@ -70,19 +68,19 @@ export const WelcomeEmail = ({ firstName }: WelcomeEmailProps) => {
                   </ol>
 
                   <Text style={paragraph}>
-                    To make sure you don't miss out, add{' '}
+                    To make sure you don&apos;t miss out, add{' '}
                     <Link
                       href="mailto:hello@mikebifulco.com"
                       className="text-pink-600"
                     >
                       hello@mikebifulco.com
                     </Link>{' '}
-                    to your contacts. And if Tiny Improvements isn't your thing,
-                    you can unsubscribe anytime — no hard feelings.
+                    to your contacts. And if Tiny Improvements isn&apos;t your
+                    thing, you can unsubscribe anytime — no hard feelings.
                   </Text>
                   <Text style={paragraph}>
-                    Thanks again for joining Tiny Improvements. I can't wait to
-                    see what you build!
+                    Thanks again for joining Tiny Improvements. I can&apos;t
+                    wait to see what you build!
                   </Text>
 
                   <Text style={paragraph}>
@@ -104,12 +102,13 @@ export const WelcomeEmail = ({ firstName }: WelcomeEmailProps) => {
               </Row>
             </Section>
 
-            <Section style={containerImageFooter}>
+            <Section style={containerImageFooter} align="center">
               <Img
-                style={image}
+                className="mx-auto max-w-full"
                 width={620}
                 height={92}
                 src={`https://res.cloudinary.com/mikebifulco-com/image/upload/v1723993697/email/footerimg.png`}
+                alt="These lovely plant graphics are licensed from Jyothi on The Noun Project: https://thenounproject.com/creator/jyothi.shilpa077"
               />
             </Section>
 
@@ -158,10 +157,6 @@ const content = {
   overflow: 'hidden',
   maxWidth: '500px',
   backgroundColor: '#fff',
-};
-
-const image = {
-  maxWidth: '100%',
 };
 
 const containerImageFooter = {
